test_server/requirements/dev.txt
# requirements/dev.txt
-r common.txt

django-debug-toolbar==1.9.1